+++ requirements/base.txt
@@ -6,7 +6,7 @@ Django==2.1
 django-embed-video==1.0.0
 djet==0.0.10
 Fabric==1.13.2
-factory-boy==2.5.2
+factory-boy==2.12.0
 gunicorn==19.7.1
 idna==2.5
 packaging==16.8
